backend/requirements.txt
opengradient>=0.3.0
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.0
pydantic>=2.0.0
web3>=6.0.0
httpx>=0.27.0
